feat: add feature spreads units location and searchbox, fix bug next-auth error building app
