+++ package.json
@@ -21,6 +21,7 @@
     "@radix-ui/react-slot": "^1.2.3",
     "@radix-ui/react-switch": "^1.2.5",
     "@tanstack/react-query": "^5.77.2",
+    "@tanstack/react-query-devtools": "^5.83.0",
     "@vis.gl/react-google-maps": "^1.5.3",
     "axios": "^1.9.0",
     "class-variance-authority": "^0.7.1",
@@ -30,7 +31,7 @@
     "lucide-react": "^0.511.0",
     "motion": "^12.16.0",
     "next": "15.1.8",
-    "next-auth": "^4.24.11",
+    "next-auth": "^5.0.0-beta.29",
     "next-themes": "^0.4.6",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
