+++ src/lib/axios-instance.ts
@@ -1,13 +1,11 @@
 import axios from "axios";
 
+const url = process.env.BACKEND_URL;
+
 const axiosInstance = axios.create({
-  baseURL: process.env.BACKEND_URL,
+  baseURL: url,
   headers: {
     "Content-Type": "application/json",
-    Connection: "keep-alive",
-  },
-  transitional: {
-    clarifyTimeoutError: true,
   },
 });
 
@@ -16,7 +14,7 @@ axiosInstance.interceptors.response.use(
   (error) => {
     if (error.code === "ECONNABORTED") {
       console.error("Timeout - Increase backend speed or timeout");
-    } else if (error.message.includes("stream has aborted")) {
+    } else if (error.message.includes("stream has been aborted")) {
       console.error("Stream killed - Check network/memory limits");
     }
     return Promise.reject(error);
